chore(IT Wallet): [SIW-586] Refactor navigation routes (#5244)
## Short description
Depends on #5241.
This PR suggests a refactoring of the navigation routes, introducing a
structure which separates the issuance and presentation flows for PID
and generic credentials. While the ultimate aim is to merge these flows
in the future, the current proposal might be beneficial as well.
Subsequently, another PR will follow to align the screens folder
structure with the newly proposed routes structure.

## List of changes proposed in this pull request
- Updates the routes structure;
- Updates routes names;
- Updates usage.

## How to test
Static tests. Optionally check every flow again.