chore(IT Wallet): [SIW-693,SIW-691] Add released by and issued by claims bottom sheet (#5248)
## Short description
This PR introduces the bottom sheet for the `issued by` and `released
by` claims in the preview and details screen of a credential. It also
removes it from the `assistance` banner which now navigates to the not
available screen.

## List of changes proposed in this pull request
- Adds two bottom sheets for each claim;
- Removes the bottom sheet from the `assistance` banner and adds a
navigation the not available screen.

## How to test
Try to obtain a new credential and test the bottom sheets for the two
claims in the preview screen. Also test the navigation when pressing on
the assistance banner. Then test them in the credential preview screen
by single pressing it in the wallet home screen.