chore(IT Wallet): [SIW-675] Add QR Code screen during proximity flow (#5238)
## Short description
This PR adds the QRCode screen during proximity flow (for testing
purposes the qr-code content is mocked with a fake mDoc payload). Having
created the proximity package that uses an event manager on the UX/UI
side, messages can be conveyed based on the progress of the
presentation. I therefore used a loading message, a message that can
vary based on the progress of the presentation.


https://github.com/pagopa/io-app/assets/49342144/f55ffed6-0562-4bf1-880c-da7efc72eb2d

## List of changes proposed in this pull request
- Add ItwPresentationQrCodeScreen.tsx

## How to test
Open the credential details and tap "Show QR Code". The new screen
should appear.

---------

Co-authored-by: LazyAfternoons <LazyAfternoons@users.noreply.github.com>