fix(IT Wallet): [SIW-570] Blocking error screen in wallet home screen (#5259)
## Short description
This PR fixes a blocking error screen in the wallet home screen.
It's shown when the decode of the PID fails. That's not supposed to
happen but sometimes it does due to dirty state which we are currently
not migrating.

## List of changes proposed in this pull request
- Replaces the old error view with a new one which has a reset button
and doesn't block the wallet usage.

## How to test
Static checks should be enough.

Co-authored-by: Mario Perrotta <mario.perrotta@pagopa.it>