feat: [IOPID-689] Email validation at startup (#5264)
## ⚠️ Depends on https://github.com/pagopa/io-app/pull/5206 ⚠️ ##

## Short description
This PR blocks the user if their email is not verified or is already
taken, forcing them to validate or change it. The flow is described in
this
[Figma](https://www.figma.com/file/oLJMcWd0SbRHDtj7WRvu7N/Univocit%C3%A0-indirizzi-email?type=design&node-id=31-12161&mode=design&t=np5Q8HLkqXhrq6ey-4).

## List of changes proposed in this pull request
- The two modal `ValidateEmailModal` and `EmailAlreadyUsedModal` has
been changed to their sibling screen component `ValidateEmailScreen` and
`EmailAlreadyTakenScreen` to use the navigation capabilities.
- A new redux state has been added to `EmailValidationState` to know if
we are landing to `EmailInsertScreen` from the standard flow or because
the user has been block during the startup process.

## How to test
Enable the `new CDU flow` and using the `io-dev-api-server` try to:
- Enter the app as a new user (firstOnboardig: true) and see that you
are blocked until you have validated your email. Even if you reload the
app.
- Change your email in the profile section and see that you are blocked
until you have validated your email, even if you reload the app.
- Try to invalidate the email and set it as "already taken", and restart
the app. You should see that you are blocked until you have changed and
successfully validated the new email, even if you reload the app.

> At app startup if `isEmailValidated=true` the user can continue
without problems, otherwise if `isEmailValidated=false` the user sees
`ValidateEmailScreen` or `EmailAlreadyTakenScreen` based on the value of
`isEmailAlreadyTaken`.

<details open><summary>Details</summary>
<p>

<video
src="https://github.com/pagopa/io-app/assets/16268789/6aa4156e-71e0-4cc1-9071-19fafbe8bbc8"
/>

</p>
</details>

---------

Co-authored-by: Alice Di Rico <83651704+Ladirico@users.noreply.github.com>
Co-authored-by: Alice Azzolini <93777761+AliceAzzolini@users.noreply.github.com>
Co-authored-by: Sabino <114305019+sabontech@users.noreply.github.com>
Co-authored-by: Cristiano Tofani <cri.tofani@gmail.com>
Co-authored-by: Andrea Piai <andrea.piai@pagopa.it>
Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>
Co-authored-by: Alessandro Izzo <34343582+Hantex9@users.noreply.github.com>