feat: [IOCOM-695] Add the `REMOTE_CONTENT_INFO` mixpanel event (#5309)
This PR depends on https://github.com/pagopa/io-app/pull/5298

## Short description
This PR adds support for `REMOTE_CONTENT_INFO` mixpanel event. This
event is triggered when the user taps on the link in the
RemoteContentBanner.

## List of changes proposed in this pull request
See commit list

## How to test
Using `io-dev-api-server` generate a message with remote content. Open
the message detail, tap on the link in the banner.