chore(IT Wallet): [SIW-722] Add verifyAndParseCredential phase to credential issuing (#5310)
## Short description
Depends on
[this](https://github.com/pagopa/io-react-native-wallet/pull/71) PR.

This PR introduces the `verifyAndParseCredential` phase to the
credential issuing flow.

## List of changes proposed in this pull request
- Replaces the manual verify and parse phase in `ItwIssuanceSaga.ts`
with the library `verifyAndParseCredential` function which already
returns a Record of claims;
- Adjusts `ItwCredentialClaimsList.tsx` to get the claims from the new
Record structure. There's no need to use the
`credentialConfigurationSchema` anymore.

## How to test
Reset the wallet. Test a credential issuing flow, the claims should be
displayed just as before.

---------

Co-authored-by: Mario Perrotta <mario.perrotta@pagopa.it>