chore(IT Wallet): [SIW-789] Refactor folders and files structure (#5384)
## Short description
Depends on #5373 
This PR focuses on enhancing the organization and coherence of the
project's folder and file structure. No changes to the logic have been
made whatsoever.
While a detailed breakdown of each modification isn't provided due to
the high number of changes. They can be reviewed in the `Files changed`
section or by pulling the branch locally to inspect the updated folder
structure and file names.

It also adds a state migration needed due to the changes which basically
resets it.

## How to test
Check the new structure and names statically then test again each flow.
Everything should work just as before.