feat: [IOCOM-848] Message content with new DS System (#5519)
This PR depends on #5516 

## Short description
This PR brings the new DS to the message content.

## List of changes proposed in this pull request
- LegacyMarkdown to avoid disruptions in other flows
- New Markdown component which is a refactoring of the old class one to
a functional one, with new DS
- Message content is rendered using the new Markdown component

Please take note that the old markdown component remains in place and it
is used widely throughout screens. It has not been replaced because of
the refactoring from a class component to a functional one, merging
three lifecycle methods (mount, update, unmount) into one hook
(useEffect). In the new component, this may trigger an unwanted
recompiling of the markdown if the caller component does not handle the
input properties accordingly. So, to avoid massive disruption, the old
class component lives and before replacing it, one should carefully
check for unwanted behaviours due to bad input handling.

## How to test
Enable the new DS and check that the message content is rendered
properly.

---------

Co-authored-by: Alessandro Dell'Oste <alessandro.delloste1@gmail.com>