chore: [IOPID-1481], [IOPID-1482] Countdown 60 seconds (#5514)
## Short description
I implemented the intermediate solution for this screen. 
We have maintained the as-is logic and have added a timer that notifies
the user when he/she can try again to get the email sent.

## List of changes proposed in this pull request
- Fix logic of email validation email send. We have changed the cases
where the email is sent by the FE. The email is only sent when we are on
the validation screen and the email is the same, when the email is
different it is automatically sent by the BE
- Add logic to show the countdown when send button is disabled
- The timeout for the user to activate the re-send button was increased
(from 10 seconds to 60 seconds)
- Fix tests
- change .env.local configuration to make it compliant with
.env.production



https://github.com/pagopa/io-app/assets/83651704/de83b202-7802-4cd6-acef-25bdb6a60581



## How to test
Run the application and run the validation email flow. (or
onboarding/first onboarding or by profile section)

---------

Co-authored-by: Fabio Bombardi <16268789+shadowsheep1@users.noreply.github.com>