feat: [IOCOM-1207] Payment data on the "More info" bottom sheet (#5695)
## Short description
This PR adds the payment data (notice number and payee fiscal code) to
the "More info" bottom sheet, in the message details screen, with the
new DS enabled.

|No payment data|With payment data|
|-|-|
|![Simulator Screenshot - iPhone 15 - 2024-04-16 at 11 53
59](https://github.com/pagopa/io-app/assets/5150343/08e9d1b8-b512-4d0d-8e08-dc4cccebbe7a)|![Simulator
Screenshot - iPhone 15 - 2024-04-16 at 11 49
25](https://github.com/pagopa/io-app/assets/5150343/fad8b22c-1874-4a1a-97a8-0cefb174d452)|

## List of changes proposed in this pull request
- `ShowMoreListItem` supports the visualisation of payment's notice code
and payee's fiscal code.
- New `formatPaymentNoticeNumber` function, in order to share common
formatting logic for a notice number

## How to test
Using the io-dev-api-server, generate a message without a payment and a
message with a payment. Check that the former does not show any payment
info while the latter does.